添加bloomFilter去重以及相关的用例 修复urlPool中totalCount和leftCount不正确的bug

+++ src/main/java/com/myway5/www/Urlpool/Remover/HashMapDuplicateRemover.java
@@ -13,7 +13,7 @@ public class HashMapDuplicateRemover implements IDuplicateUrlRemover{
 		return true;
 	}
 
-	public int getTotalUrlCount() {
+	public long getTotalUrlCount() {
 		return urls.size();
 	}
